deleting an image with ajax works

+++ app/Http/Controllers/ImageController.php
@@ -10,7 +10,7 @@ use App\Http\Requests\ImageRequest;
 class ImageController extends Controller
 {
     public function __construct() {
-        $this->middleware('auth', ['only' => ['edit', 'destroy']]);
+        $this->middleware('isAdmin', ['only' => ['edit', 'destroy']]);
     }
     public function index() {
         $images = Image::all();
@@ -58,9 +58,10 @@ class ImageController extends Controller
         $image->tags()->sync($request->tags);
         return  redirect('gallery');
     }
-    public function destroy($image_id) {
-        $image = Image::findOrFail($image_id);
-        $image->delete();
-        return redirect('gallery');
+    public function destroy(Request $request) {
+        if($request->ajax()) {
+            $image = Image::findOrFail($request->image_id);
+            $image->delete();
+        }
     }
 }
